Object Pool에 의해 관리되는 Projectile을 나타내는 AAuraProjectile 하위 클래스를 생성하고 기존 Pooled Projectile의 부모 클래스로 설정 - EPooledActorType 프로퍼티 추가 기존 Firebolt_Pooled를 제거하고 FireBolt로 통합
