+++ Source/Aura/Private/Actor/AuraProjectile.cpp
@@ -10,8 +10,6 @@
 #include "Components/AudioComponent.h"
 #include "Components/SphereComponent.h"
 #include "GameFramework/ProjectileMovementComponent.h"
-#include "Kismet/GameplayStatics.h"
-
 
 AAuraProjectile::AAuraProjectile()
 {
