+++ Source/Aura/Private/Actor/FireBolt.cpp
@@ -3,6 +3,7 @@
 
 #include "Actor/FireBolt.h"
 
+#include "AbilitySystemBlueprintLibrary.h"
 #include "AuraGameplayTags.h"
 #include "NiagaraComponent.h"
 
@@ -10,9 +11,21 @@ AFireBolt::AFireBolt()
 {
 	FireEffectComponent = CreateDefaultSubobject<UNiagaraComponent>(TEXT("Fire Effect Component"));
 	FireEffectComponent->SetupAttachment(GetRootComponent());
-
 	ImpactCueTag = AuraGameplayTags::GameplayCue_FireBoltImpact;
+
+	PooledActorType = EPooledActorType::FireBolt;
 }
 
+void AFireBolt::SetInUse(bool bInUse)
+{
+	Super::SetInUse(bInUse);
 
-
+	if (bInUse)
+	{
+		FireEffectComponent->Activate();
+	}
+	else
+	{
+		FireEffectComponent->Deactivate();
+	}
+}
