+++ Source/Aura/Public/Actor/FireBall.h
@@ -3,31 +3,27 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "AuraProjectile.h"
+#include "PooledProjectile.h"
 #include "FireBall.generated.h"
 
-class UPooledActorComponent;
 class UNiagaraComponent;
 
 /**
  * 폭발 후 사방으로 EmberBolt를 발사하는 FireBall Projectile
- * Object Pool에 의해 관리된다.
  */
 UCLASS()
-class AURA_API AFireBall : public AAuraProjectile
+class AURA_API AFireBall : public APooledProjectile
 {
 	GENERATED_BODY()
 
 public:
 	AFireBall();
-	virtual void PostInitializeComponents() override;
 
 	// FireBall 폭발 후 EmberBolt를 Spawn하기 위해 옵션 설정
 	void SetEmberBoltOptions(int32 InNumEmberBolts, float InEmberBoltDamage, const TSubclassOf<AAuraProjectile>& InEmberBoltClass);
 
 protected:
 	virtual void OnSphereOverlap(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult) override;
-	virtual bool IsValidOverlap(const AActor* TargetActor) const override;
 
 private:
 	UPROPERTY(VisibleAnywhere)
@@ -39,8 +35,5 @@ private:
 
 	void SpawnEmberBolts() const;
 
-	UPROPERTY(VisibleAnywhere)
-	TObjectPtr<UPooledActorComponent> PooledActorComponent;
-
-	void SetInUse(bool bInUse) const;
+	virtual void SetInUse(bool bInUse) override;
 };
